sun8i-emac: Do not generate random hardware address in sun8i_emac_set_macaddr
sun8i_emac_set_macaddr() sets MAC addresses a given slot. Slot 0 is the
controller's address. Slot 1+ are for hash filters.

The function should not care whether the address is valid. It is the
caller's job to be sure. For slot 0 we already generate a random address
when there's no address provided in the DT. Also, re-generating the
address stored in netdev, while not updating the hardware values is
not correct either.

Also no need to set the MAC address at probe time. It's probably going
to get cleared when we do a soft reset when the interface is opened. It
is set again afterwards.